refactor: Organize trigger tests into focused test modules
Split the monolithic 1,200-line trigger_tests.rs into focused modules:
- ddl.rs: CREATE/DROP trigger tests
- insert.rs: INSERT trigger firing tests
- update.rs: UPDATE trigger firing tests
- delete.rs: DELETE trigger firing tests
- coordination.rs: Multiple triggers, ordering tests
- conditions.rs: WHEN clause filtering tests
- error_handling.rs: Failures, rollback, recursion tests
- pseudo_vars.rs: OLD/NEW variable tests
- mod.rs: Shared helpers and module exports

Benefits:
- Logical organization: Related tests grouped together
- Easier navigation: Find INSERT tests in insert.rs, etc.
- Better discoverability: Clear test coverage structure
- Reduced file size: Each module ~100-200 lines instead of 1,200
- Improved maintainability: Changes isolated to relevant modules

All 21 tests pass without modification.

Closes #2486
